Breaks NetworkInfo out into suitable variants for IAAS and CAAS implementations.
